sing-language-backend-python/requirements.txt
flask
flask-socketio
eventlet
tensorflow-macos
tensorflow-metal
numpy
pillow
